Fashion photo app
I've created an app for fashion loves. In this app people can share their styles, thoughts and views. It is a way of showing  new trends to other people and everyone has right to share their photos wearing what they think is cool and showing other people whats to wear to look good